backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
supabase==2.4.5
python-dotenv==1.0.1
